feat: routes can define their cost
