+++ packages/api/src/utils/request.ts
@@ -4,6 +4,7 @@ declare global {
 	namespace Express {
 		interface Request {
 			accessLevel?: number
+			cost?: number
 		}
 	}
 }
